Moved editString() to /common/utils.js
